% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shkr.R
\name{load_shkr2010}
\alias{load_shkr2010}
\title{Title load_shkr2010 Load shkr data}
\usage{
load_shkr2010(path = "../2data/shkr2010/")
}
\arguments{
\item{path}{Path to the downloades and unziped shkr folder}
}
\value{
a list of tibbles of the different data.
}
\description{
The function load_shkr2010 loads shkr data from
the downloades shkr 2010 data set. It is required
to downloads the data from https://www.jma.uni-kiel.de/en/research-projects/data-exchange-platform/shkr
and unzip them.
}
\examples{
shkr <- load_shkr2010()
}
\author{
Oliver Nakoinz <oliver.nakoinz@ufg.uni-kiel.de>
}
