% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shkr.R
\name{normalize}
\alias{normalize}
\title{helper-function to normalize divides the values in a row by the sum of the values}
\usage{
normalize(a)
}
\arguments{
\item{a}{tibble or dataframe}
}
\value{
the same dataframe with normalized values
}
\description{
helper-function to normalize divides the values in a row by the sum of the values
}
\examples{

}
\author{
Oliver Nakoinz <oliver.nakoinz@ufg.uni-kiel.de>
}
