% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shkr.R
\name{shkr2010_download}
\alias{shkr2010_download}
\title{Title shkr2010_download downloads the shkr database from 2010 from the JMA data exchange platform}
\usage{
shkr2010_download(path = "../2data")
}
\arguments{
\item{path}{path to the folder were the shkr db has to be stored}
}
\description{
Title shkr2010_download downloads the shkr database from 2010 from the JMA data exchange platform
}
\examples{
shkr2010_download()
}
\author{
Oliver Nakoinz <oliver.nakoinz@ufg.uni-kiel.de>
}
