% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shkr.R
\name{shkr_filter_dat}
\alias{shkr_filter_dat}
\title{\code{shkr_filter_dat} filter a shkr tibble/dataframe to a certain dating}
\usage{
shkr_filter_dat(data, dating, range = TRUE)
}
\arguments{
\item{data}{shkr tibble as produced by \code{shkr_comp_loc10}}

\item{dating}{filter type: "dating", numerical according to thesaurus}

\item{range}{logical, TRUE: dating names all desired dates (default), FALSE: names undesired names}
}
\value{
a tibble as used as input but restricted to a certain dating
}
\description{
\code{shkr_filter_dat} filter a shkr tibble/dataframe to a certain dating
}
\author{
Franziska Faupel <\email{ffaupel@ufg.uni-kiel.de}>
}
