% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shkr.R
\name{shkr_filter_loc06}
\alias{shkr_filter_loc06}
\title{\code{shkr_filter_loc06} filters certain types of sites}
\usage{
shkr_filter_loc06(data, type)
}
\arguments{
\item{data}{list of shkr tibbles as produced by \code{load_shkr2010}}

\item{type}{numeric vector of types of sites (see thesaurus above)}
}
\value{
tibble/dataframe containing the filterd loc_06 information
}
\description{
The function shkr_filter_loc06 filters certain types of sites.
                     1 = archaeolgical object,
                     2 = depot,
                     3 = singular object,
                     4 = grave
                     5 = grvae/settlement/production
                     6 = infrastructure
                     7 = cult or religion
                     8 = settlement/production
                     9 = settlement/production/cult or religion
                     10 = others
                     12 = unknown
}
\examples{
type <- 4
shkr_grave <- shkr_filter_loc06(shkr, type)

type <- c(5,8,9)
shkr_sett <- shkr_filter_loc06(shkr, type)

}
\author{
Franziska Faupel <\email{ffaupel@ufg.uni-kiel.de}>
}
