% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shkr.R
\name{shkr_filter_loc10}
\alias{shkr_filter_loc10}
\title{\code{shkr_filter_loc10} filters certain types of artefacts}
\usage{
shkr_filter_loc10(data, a = "A", b = "B", c = "C", d = "D")
}
\arguments{
\item{data}{list of shkr tibbles as produced by \code{load_shkr2010}}

\item{a}{text string of material type to filter, starts with the letter A}

\item{b}{text string of shape type to filter, starts with the letter b}

\item{ctext}{string of decoration technique type to filter, starts with the letter c}

\item{dtext}{string of decoration element type to filter, starts with the letter d}
}
\value{
tibble/dataframe containing the filterd loc_10 information
}
\description{
The function shkr_filter_loc10 filters certain artefact types
from the table of artefacts. SHRK uses a facet classification.
Facet a contains material, facet b shape, facet c decoration technique
and facet d decoration elements. Insite the facets a
hierarchical classfication is used. The classification includes
isomorph branches which have the same next feature at the same
place in different branches. The wildcard "." can be used to
filter for the same feature in different branches.
}
\examples{
a <- "A"
b <- "B31"
c <- "C"
d <- "D"
shkr_fibulae <- shkr_filter_loc10(shkr, a, b, c, d)
}
\author{
Oliver Nakoinz <oliver.nakoinz@ufg.uni-kiel.de>
}
