% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shkr.R
\name{shkr_filter_region}
\alias{shkr_filter_region}
\title{Titleshkr_filter_region filter a shkr tibble/dataframe to a certain spatial extent}
\usage{
shkr_filter_region(data, ftype = "loc01", a = 11, x_range = c(0,
  3597713), y_range = c(0, 5664752))
}
\arguments{
\item{data}{shkr tibble as produced by shkr_comp_loc10}

\item{ftype}{filter type: "coords", "loc01", "loc02", "loc03", "loc04", "loc05", "loc06"}

\item{a}{integer loc_xx_id number}

\item{x_range}{vector: c(minimal, maximal)}

\item{y_range}{vector: c(minimal, maximal)}
}
\value{
a tibble as used as input but restricted to a certain area
}
\description{
Titleshkr_filter_region filter a shkr tibble/dataframe to a certain spatial extent
}
\examples{
shkr_fibulae_comp_coord <- shkr_filter_region(shkr_fibulae_comp_coord)
}
\author{
Oliver Nakoinz <oliver.nakoinz@ufg.uni-kiel.de>
}
