% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shkr.R
\name{shkr_net}
\alias{shkr_net}
\title{\code{shkr_net} creates igraph network from distance matrix}
\usage{
shkr_net(dm)
}
\arguments{
\item{dm}{distance matric as produced by \code{shkr_cul_dist}}
}
\value{
igraph object
}
\description{
\code{shkr_net} creates igraph network from distance matrix
}
\examples{
snet <- shkr_net(dm)
}
\author{
Oliver Nakoinz <oliver.nakoinz@ufg.uni-kiel.de>
}
